Refactor test structure by moving test files to the src directory and updating preload path in bunfig.toml
