+++ src/common/timers.ts
@@ -1,28 +1,7 @@
+import { setTimeout } from "node:timers/promises";
+
 interface Options {
 	signal?: AbortSignal;
-}
-
-/**
- * Wait for a specified amount of time, accepts a signal to abort the timer.
- * @param ms The amount of time to wait in milliseconds
- * @param options The options for the timer
- * @example
- * let controller = new AbortController();
- * await wait(1000, { signal: controller.signal });
- */
-export function wait(ms: number, options?: Options): Promise<void> {
-	return new Promise((resolve, reject) => {
-		let timeout = setTimeout(() => {
-			if (options?.signal?.aborted) return reject(new TimersError("Aborted"));
-			return resolve();
-		}, ms);
-		if (options?.signal) {
-			options.signal.addEventListener("abort", () => {
-				clearTimeout(timeout);
-				reject(new TimersError("Aborted"));
-			});
-		}
-	});
 }
 
 /**
@@ -40,7 +19,7 @@ export async function* interval(ms: number, options?: Options) {
 	let signal = options?.signal ?? new AbortSignal();
 	while (!signal.aborted) {
 		try {
-			yield await wait(ms, { signal });
+			yield await setTimeout(ms, void 0, { signal });
 		} catch {
 			return;
 		}
