+++ src/server/cors.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from "bun:test";
 
-import { cors } from "../../src/server/cors";
+import { cors } from "./cors";
 
 function createRequest(headers: Headers = new Headers()) {
 	headers.set("Origin", "http://remix.utils");
