+++ src/server/get-client-id-address.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test } from "bun:test";
-import { getClientIPAddress } from "../../src/server/get-client-ip-address";
+import { getClientIPAddress } from "./get-client-ip-address";
 
 const VALID_IP = "192.168.0.1";
 const INVALID_IP = "abc.def.ghi.jkl";
