+++ src/server/get-client-locales.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test } from "bun:test";
-import { getClientLocales } from "../../src/server/get-client-locales";
+import { getClientLocales } from "./get-client-locales";
 
 describe(getClientLocales, () => {
 	test("should return undefined if no Accept-Language header is not present", () => {
