+++ src/server/preload-route-assets.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from "bun:test";
 import type { EntryContext } from "react-router";
-import { preloadRouteAssets } from "../../src/server/preload-route-assets";
+import { preloadRouteAssets } from "./preload-route-assets";
 // TODO Fix this
 describe(preloadRouteAssets.name, () => {
 	let context: EntryContext = {
