+++ src/test/fake-match.ts
@@ -1,4 +1,4 @@
-import { RouteMatch } from "react-router";
+import type { RouteMatch } from "react-router";
 
 // biome-ignore lint/suspicious/noExplicitAny: RouteData is not exported anymore, route match has issues
 export function fakeMatch(data?: any): any {
